move all the plugin code to __init__.py
